# backend/docs/swagger.yaml
definitions:
  types.LoginAdmin:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  types.NotificationRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
      type:
        type: string
    type: object
  types.RegisterUserRequest:
    properties:
      telegramUserId:
        type: integer
      username:
        type: string
    required:
    - telegramUserId
    - username
    type: object
  types.Token:
    properties:
      token:
        type: string
    required:
    - token
    type: object
info:
  contact: {}
paths:
  /admin/api/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.LoginAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
      summary: Admin login
      tags:
      - admin
  /api/notification:
    post:
      consumes:
      - application/json
      description: Sends a notification message to all authenticated users via Redis
        Streams
      parameters:
      - description: Notification
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.NotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Send notification to all users
      tags:
      - notifications
  /api/token:
    post:
      consumes:
      - application/json
      description: Validate tg token, register user if new, and return login result
      parameters:
      - description: Token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.Token'
      produces:
      - application/json
      responses:
        "200":
          description: Returns user info with isNewUser and registered flags
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Takes Telegram token
      tags:
      - auth
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Registers a Telegram user in the database
      parameters:
      - description: User registration data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - users
  /api/users/authorized:
    get:
      description: Returns a list of all authorized Telegram users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all authorized users
      tags:
      - users
swagger: "2.0"